use nx package to read nxjson
