+++ libs/vscode/nx-workspace/src/index.ts
@@ -2,3 +2,5 @@ export * from './lib/find-workspace-json-target';
 export * from './lib/reveal-workspace-json';
 export * from './lib/workspace-codelens-provider';
 export * from './lib/verify-workspace';
+export * from './lib/get-nx-config';
+export * from './lib/get-nx-workspace-config';
