+++ libs/vscode/nx-workspace/src/lib/get-nx-config.ts
@@ -1,0 +1,21 @@
+import { cacheJson, readAndCacheJsonFile } from '@nx-console/server';
+import { NxJsonConfiguration } from '@nrwl/devkit';
+import { join } from 'path';
+import { getNxWorkspacePackageFileUtils } from './get-nx-workspace-package';
+
+export function getNxConfig(baseDir: string): NxJsonConfiguration {
+  try {
+    let cachedNxJson = cacheJson('nx.json', baseDir).json;
+
+    if (!cachedNxJson) {
+      const nxJson = (getNxWorkspacePackageFileUtils() as any).readNxJson(
+        join(baseDir, 'nx.json')
+      );
+
+      cachedNxJson = cacheJson('nx.json', baseDir, nxJson).json;
+    }
+    return cachedNxJson;
+  } catch (e) {
+    return readAndCacheJsonFile('nx.json', baseDir).json;
+  }
+}
